Javadoc all classes except command classes. Clean up code.

+++ src/main/java/net/farlands/sanctuary/util/TimeInterval.java
@@ -1,5 +1,8 @@
 package net.farlands.sanctuary.util;
 
+/**
+ * Time conversion methods.
+ */
 public enum TimeInterval {
     SECOND(1, "second", "s"),
     MINUTE(60, "minute", "m"),
@@ -35,29 +38,29 @@ public enum TimeInterval {
 
     public static String formatTime(long millis, boolean abbreviate, TimeInterval minInterval) {
         long seconds = millis / 1000L;
-        if(Math.round(seconds) < 1.0)
+        if (Math.round(seconds) < 1.0)
             return "Now";
         StringBuilder sb = new StringBuilder(0);
         TimeInterval currentInterval;
-        for(int i = VALUES.length - 1;i >= minInterval.ordinal();-- i) {
+        for (int i = VALUES.length - 1;i >= minInterval.ordinal();-- i) {
             currentInterval = VALUES[i];
             int time = (int)(seconds / currentInterval.getToSecondsFactor());
             seconds %= currentInterval.getToSecondsFactor();
-            if(time == 0 && minInterval.equals(currentInterval) && !minInterval.equals(VALUES[0]) && sb.length() == 0)
+            if (time == 0 && minInterval.equals(currentInterval) && !minInterval.equals(VALUES[0]) && sb.length() == 0)
                 time = 1;
-            if(time > 0) {
-                if(sb.length() > 0) {
-                    if(!abbreviate && i == minInterval.ordinal())
+            if (time > 0) {
+                if (sb.length() > 0) {
+                    if (!abbreviate && i == minInterval.ordinal())
                         sb.append(" and ");
                     else
                         sb.append(' ');
                 }
                 sb.append(time);
-                if(abbreviate)
+                if (abbreviate)
                     sb.append(currentInterval.getAbbreviation());
-                else{
+                else {
                     sb.append(' ').append(currentInterval.getName());
-                    if(time > 1)
+                    if (time > 1)
                         sb.append('s');
                 }
             }
@@ -70,27 +73,27 @@ public enum TimeInterval {
     }
 
     public static long parseSeconds(String time) {
-        if(time == null || time.isEmpty())
+        if (time == null || time.isEmpty())
             return -1L;
-        if("now".equalsIgnoreCase(time))
+        if ("now".equalsIgnoreCase(time))
             return 0L;
         long totalSeconds = 0L;
         String[] timeElements = time.split(" ");
-        for(String te : timeElements) {
+        for (String te : timeElements) {
             TimeInterval ti = null;
-            for(TimeInterval ti0 : VALUES) {
-                if(te.endsWith(ti0.getAbbreviation()))
+            for (TimeInterval ti0 : VALUES) {
+                if (te.endsWith(ti0.getAbbreviation()))
                     ti = ti0;
             }
-            if(ti == null)
+            if (ti == null)
                 return -1L;
             int t;
             try {
                 t = Integer.parseInt(te.substring(0, te.length() - ti.getAbbreviation().length()));
-            }catch(NumberFormatException ex) {
+            } catch (NumberFormatException ex) {
                 return -1L;
             }
-            totalSeconds += t * ti.getToSecondsFactor();
+            totalSeconds += (long) t * ti.getToSecondsFactor();
         }
         return totalSeconds;
     }
